Add a base class for the device classes to avoid code duplication
JSRemoteTest-DCO-1.0-Signed-off-by: Peter Marki marpeter@inf.u-szeged.hu
