+++ API/testrunner/devices/artik053.py
@@ -14,22 +14,20 @@
 
 import time
 
+from API.testrunner.devices.device_base import RemoteDevice
 from API.common import console, paths, utils
 from API.testrunner.devices.connections.serialcom import SerialConnection
 from threading import Thread
 
-class ARTIK053Device(object):
+class ARTIK053Device(RemoteDevice):
     '''
     Device of the ARTIK053 target.
     '''
     def __init__(self, env):
         self.os = 'tizenrt'
-        self.app = env['info']['app']
         self.tizenrt = env['modules']['tizenrt']
-        self.env = env
 
-        # Check the members before testing.
-        self.check_args()
+        RemoteDevice.__init__(self, env)
 
         data = {
             'dev-id': env['info']['device_id'],
@@ -39,13 +37,6 @@ class ARTIK053Device(object):
         }
 
         self.channel = SerialConnection(data)
-
-    def check_args(self):
-        '''
-        Check that all the arguments are established.
-        '''
-        if not self.env['info']['device_id']:
-            console.fail('Please use the --device-id to select the device.')
 
     def initialize(self):
         '''
@@ -66,37 +57,6 @@ class ARTIK053Device(object):
         utils.execute(self.tizenrt['paths']['os'], 'make', flags, quiet=True)
         # Wait a moment to boot the device.
         time.sleep(2)
-
-    def login(self):
-        '''
-        Login to the device.
-        '''
-        try:
-            self.channel.open()
-
-            # Press enters to start the serial communication and
-            # go to the test folder because some tests require resources.
-            self.channel.exec_command('\n\n')
-            self.channel.exec_command('cd /rom')
-
-            if self.env['info']['coverage'] and self.app == 'iotjs':
-                # Set up the wifi connection.
-                wifi_name = utils.get_environment('ARTIK_COV_WIFI_NAME')
-                wifi_pwd =  utils.get_environment('ARTIK_COV_WIFI_PWD')
-
-                self.channel.exec_command('wifi startsta')
-                self.channel.exec_command('wifi join %s %s' % (wifi_name, wifi_pwd))
-                self.channel.exec_command('ifconfig wl1 dhcp')
-                time.sleep(1)
-
-        except Exception as e:
-            console.fail(str(e))
-
-    def logout(self):
-        '''
-        Logout from the device.
-        '''
-        self.channel.close()
 
     def execute(self, testset, test):
         '''
