+++ API/testrunner/devices/stm32f4dis.py
@@ -15,21 +15,18 @@
 import time
 
 from API.common import console, utils
+from API.testrunner.devices.device_base import RemoteDevice
 from API.testrunner.devices.connections.serialcom import SerialConnection
 
-
-class STM32F4Device(object):
+class STM32F4Device(RemoteDevice):
     '''
     Device of the STM32F4-Discovery target.
     '''
     def __init__(self, env):
         self.os = 'nuttx'
-        self.app = env['info']['app']
         self.stlink = env['modules']['stlink']
-        self.env = env
 
-        # Check the members before testing.
-        self.check_args()
+        RemoteDevice.__init__(self, env)
 
         data = {
             'dev-id': env['info']['device_id'],
@@ -39,13 +36,6 @@ class STM32F4Device(object):
         }
 
         self.channel = SerialConnection(data)
-
-    def check_args(self):
-        '''
-        Check that all the arguments are established.
-        '''
-        if not self.env['info']['device_id']:
-            console.fail('Please use the --device-id to select the device.')
 
     def initialize(self):
         '''
@@ -70,27 +60,6 @@ class STM32F4Device(object):
         utils.execute('.', flasher, ['reset'], quiet=True)
         # Wait a moment to boot the device.
         time.sleep(5)
-
-    def login(self):
-        '''
-        Login to the device.
-        '''
-        try:
-            self.channel.open()
-
-            # Press enters to start the serial communication and
-            # go to the test folder because some tests require resources.
-            self.channel.exec_command('\n\n')
-            self.channel.exec_command('cd /test')
-
-        except Exception as e:
-            console.fail(str(e))
-
-    def logout(self):
-        '''
-        Logout from the device.
-        '''
-        self.channel.close()
 
     def execute(self, testset, test):
         '''
